only create state for used columns
